Add equeue server.
The server listen on an unix socket and wait for a json request which
look like :
  {"command": "/command/to/run",
   "timestamp": 0}

The server respond sending back the command it will run.

The server is single threaded on run each command only one by one.

If one command requested two times with the same time stamp, it
will be runned only once.