Add submanager script.
This script allow to run a command and restart it
it if some files changes.

The submanager catch SIGUSR1 and checks if the specified
files has changed.

It can use a pid file where its own pid is written.

This is usefull for restarting a service from a recipe.

Example :
  submanager -c config.cfg -- command --with-options -y
    Run “command --with-options -y” and restart it if
    config.cfg is modified after receiving SIGUSR